+++ app/models/post.rb
@@ -1,7 +1,7 @@
 class Post < ApplicationRecord
   extend FriendlyId
   friendly_id :title ,use: %i[slugged history finders]
-  include Image
+  include ImageValidation
 
   belongs_to :account
   belongs_to :community
